requirements.txt
djangorestframework
markdown
django-filter
mysqlclient
django-cors-headers
pyjwt
pillow